Fix Windows warnings exposed by /W1 accross the codebase
Also enables /W1 and /WX to ensure new W1 violations are not added. The
end goal is to enable /W3 /WX for all Windows builds.

+++ tools/oesgx/oesgx.c
@@ -152,7 +152,8 @@ int main(int argc, const char* argv[])
         epc_size =
             ((regs.ecx & 0x0fffff000) |
              ((uint64_t)(regs.edx & 0x0fffff) << 32));
-        printf("EPC size on the platform: %lu\n", epc_size);
+        printf(
+            "EPC size on the platform: %llu\n", (unsigned long long)epc_size);
     }
     return 0;
 }
